api/Food_Classifier/requirements.txt
uvicorn
numpy==1.19.1
Pillow
pathlib2
requests
matplotlib 
tensorflow==2.7.0  
scikit-learn 
fastapi
protobuf==3.20
onnxruntime
  
